refactor(cli): restructure folders - based on process(init,add...)
