+++ packages/cli/src/add/utils/copyComponent.ts
