+++ packages/cli/src/add/utils/getConfig.ts
