+++ packages/cli/src/init/index.ts
@@ -1,0 +1,6 @@
+import { checkPandaInit } from "./utils/checkPandaInit"
+import { initTemplate } from "./utils/initTemplate"
+export async function init() {
+  const root = await checkPandaInit()
+  initTemplate(root)
+}
