+++ packages/cli/src/init/utils/checkPandaInit.ts
