+++ packages/cli/src/init/utils/initTemplate.ts
