+++ packages/cli/src/init/utils/parseConfig.cjs
