+++ packages/cli/src/init/utils/template.ts
